--- a/api/types/worker.proto
+++ b/api/types/worker.proto
@@ -2,17 +2,14 @@ syntax = "proto3";
 
 package moby.buildkit.v1.types;
 
-import "github.com/gogo/protobuf/gogoproto/gogo.proto";
-import "github.com/moby/buildkit/solver/pb/ops.proto";
+option go_package = "github.com/moby/buildkit/api/types;moby_buildkit_v1_types";
 
-option (gogoproto.sizer_all) = true;
-option (gogoproto.marshaler_all) = true;
-option (gogoproto.unmarshaler_all) = true;
+import "github.com/moby/buildkit/solver/pb/ops.proto";
 
 message WorkerRecord {
 	string ID = 1;
 	map<string, string> Labels = 2;
-	repeated pb.Platform platforms = 3 [(gogoproto.nullable) = false];
+	repeated pb.Platform platforms = 3;
 	repeated GCPolicy GCPolicy = 4;
 	BuildkitVersion BuildkitVersion = 5;
 }
@@ -22,10 +19,10 @@ message GCPolicy {
 	int64 keepDuration = 2;
 	repeated string filters = 4;
 
-	int64 minStorage = 5 [(gogoproto.nullable) = true];
+	int64 minStorage = 5;
 	// maxStorage was renamed from freeBytes
-	int64 maxStorage = 3 [(gogoproto.nullable) = true];
-	int64 free = 6 [(gogoproto.nullable) = true];
+	int64 maxStorage = 3;
+	int64 free = 6;
 }
 
 message BuildkitVersion {