--- a/cache/contenthash/generate.go
+++ b/cache/contenthash/generate.go
@@ -1,3 +1,3 @@
 package contenthash
 
-//go:generate protoc -I=. -I=../../vendor/ --gogofaster_out=. checksum.proto
+//go:generate protoc -I=. -I=../../vendor/ --go_out=paths=source_relative:. checksum.proto