--- a/client/llb/marshal.go
+++ b/client/llb/marshal.go
@@ -2,34 +2,41 @@ package llb
 
 import (
 	"io"
-	"maps"
 
 	"github.com/containerd/platforms"
 	"github.com/moby/buildkit/solver/pb"
 	digest "github.com/opencontainers/go-digest"
+	"google.golang.org/protobuf/proto"
 )
 
 // Definition is the LLB definition structure with per-vertex metadata entries
 // Corresponds to the Definition structure defined in solver/pb.Definition.
 type Definition struct {
 	Def         [][]byte
-	Metadata    map[digest.Digest]pb.OpMetadata
+	Metadata    map[digest.Digest]OpMetadata
 	Source      *pb.Source
 	Constraints *Constraints
 }
 
 func (def *Definition) ToPB() *pb.Definition {
+	metas := make(map[string]*pb.OpMetadata, len(def.Metadata))
+	for dgst, meta := range def.Metadata {
+		metas[string(dgst)] = meta.ToPB()
+	}
 	return &pb.Definition{
 		Def:      def.Def,
 		Source:   def.Source,
-		Metadata: maps.Clone(def.Metadata),
+		Metadata: metas,
 	}
 }
 
 func (def *Definition) FromPB(x *pb.Definition) {
 	def.Def = x.Def
 	def.Source = x.Source
-	def.Metadata = maps.Clone(x.Metadata)
+	def.Metadata = make(map[digest.Digest]OpMetadata, len(x.Metadata))
+	for dgst, meta := range x.Metadata {
+		def.Metadata[digest.Digest(dgst)] = NewOpMetadata(meta)
+	}
 }
 
 func (def *Definition) Head() (digest.Digest, error) {
@@ -40,14 +47,14 @@ func (def *Definition) Head() (digest.Digest, error) {
 	last := def.Def[len(def.Def)-1]
 
 	var pop pb.Op
-	if err := (&pop).Unmarshal(last); err != nil {
+	if err := proto.Unmarshal(last, &pop); err != nil {
 		return "", err
 	}
 	if len(pop.Inputs) == 0 {
 		return "", nil
 	}
 
-	return pop.Inputs[0].Digest, nil
+	return digest.Digest(pop.Inputs[0].Digest), nil
 }
 
 func WriteTo(def *Definition, w io.Writer) error {
@@ -65,7 +72,7 @@ func ReadFrom(r io.Reader) (*Definition, error) {
 		return nil, err
 	}
 	var pbDef pb.Definition
-	if err := pbDef.Unmarshal(b); err != nil {
+	if err := proto.Unmarshal(b, &pbDef); err != nil {
 		return nil, err
 	}
 	var def Definition
@@ -104,7 +111,7 @@ func MarshalConstraints(base, override *Constraints) (*pb.Op, *pb.OpMetadata) {
 		Constraints: &pb.WorkerConstraints{
 			Filter: c.WorkerConstraints,
 		},
-	}, &c.Metadata
+	}, c.Metadata.ToPB()
 }
 
 type MarshalCache struct {
@@ -118,9 +125,11 @@ type MarshalCache struct {
 func (mc *MarshalCache) Cached(c *Constraints) bool {
 	return mc.dt != nil && mc.constraints == c
 }
+
 func (mc *MarshalCache) Load() (digest.Digest, []byte, *pb.OpMetadata, []*SourceLocation, error) {
 	return mc.digest, mc.dt, mc.md, mc.srcs, nil
 }
+
 func (mc *MarshalCache) Store(dt []byte, md *pb.OpMetadata, srcs []*SourceLocation, c *Constraints) {
 	mc.digest = digest.FromBytes(dt)
 	mc.dt = dt