Remove gogoproto in favor of the standard protobuf compiler. This
removes any nonstandard extensions that were part of gogoproto such as
the custom types.

Re-spin of https://github.com/moby/buildkit/pull/4422