@jsternberg 2024-09-19
This is ready for review now. This should replace https://github.com/moby/buildkit/pull/4422. I've separated the vtproto stuff into its own commit but the dependency seems to work and alleviates or removes the concern that this will affect the performance.