@crazy-max 2024-09-19
Not sure why but I see freebsd job fails: https://github.com/moby/buildkit/actions/runs/10948325927/job/30399181953?pr=5342#step:10:12

```
No usable default provider could be found for your system.

Vagrant relies on interactions with 3rd party systems, known as
"providers", to provide Vagrant with resources to run development
environments. Examples are VirtualBox, VMware, Hyper-V.

The easiest solution to this message is to install VirtualBox, which
is available for free on all major platforms.

If you believe you already have a provider available, make sure it
is properly installed and configured. You can see more details about
why a particular provider isn't working by forcing usage with
`vagrant up --provider=PROVIDER`, which should give you a more specific
error message for that particular provider.
Error: Process completed with exit code 1.
```

And same happens on fsutil repo for https://github.com/tonistiigi/fsutil/pull/206: https://github.com/tonistiigi/fsutil/actions/runs/10930826082/job/30344567135?pr=206