@jsternberg 2024-09-20
Benchmark comparisons from some areas that might be hotspots.

Comparison between gogo and the standard implementation.

```
goos: darwin
goarch: arm64
pkg: github.com/moby/buildkit/api/services/control
                         │ ../gogo.txt  │              ../std.txt              │
                         │    sec/op    │    sec/op     vs base                │
MarshalVertex-10           381.2n ± ∞ ¹   328.0n ± ∞ ¹       ~ (p=1.000 n=1) ²
MarshalVertexStatus-10     440.0n ± ∞ ¹   345.0n ± ∞ ¹       ~ (p=1.000 n=1) ²
MarshalVertexLog-10        258.5n ± ∞ ¹   239.3n ± ∞ ¹       ~ (p=1.000 n=1) ²
UnmarshalVertex-10         381.4n ± ∞ ¹   461.2n ± ∞ ¹       ~ (p=1.000 n=1) ²
UnmarshalVertexStatus-10   404.7n ± ∞ ¹   433.2n ± ∞ ¹       ~ (p=1.000 n=1) ²
UnmarshalVertexLog-10      221.5n ± ∞ ¹   284.2n ± ∞ ¹       ~ (p=1.000 n=1) ²
geomean                    337.7n         339.7n        +0.61%
¹ need >= 6 samples for confidence interval at level 0.95
² need >= 4 samples to detect a difference at alpha level 0.05

pkg: github.com/moby/buildkit/cache/contenthash
                         │ ../gogo.txt  │               ../std.txt               │
                         │    sec/op    │    sec/op      vs base                 │
MarshalCacheRecords-10     341.0n ± ∞ ¹    611.2n ± ∞ ¹        ~ (p=1.000 n=1) ²
UnmarshalCacheRecords-10   899.3n ± ∞ ¹   1274.0n ± ∞ ¹        ~ (p=1.000 n=1) ²
geomean                    553.8n          882.4n        +59.35%
¹ need >= 6 samples for confidence interval at level 0.95
² need >= 4 samples to detect a difference at alpha level 0.05
```

Comparison between gogo and vtproto.

```
goos: darwin
goarch: arm64
pkg: github.com/moby/buildkit/api/services/control
                         │ ../gogo.txt  │            ../vtproto.txt             │
                         │    sec/op    │    sec/op     vs base                 │
MarshalVertex-10           381.2n ± ∞ ¹   202.5n ± ∞ ¹        ~ (p=1.000 n=1) ²
MarshalVertexStatus-10     440.0n ± ∞ ¹   192.3n ± ∞ ¹        ~ (p=1.000 n=1) ²
MarshalVertexLog-10        258.5n ± ∞ ¹   166.8n ± ∞ ¹        ~ (p=1.000 n=1) ²
UnmarshalVertex-10         381.4n ± ∞ ¹   388.2n ± ∞ ¹        ~ (p=1.000 n=1) ²
UnmarshalVertexStatus-10   404.7n ± ∞ ¹   245.2n ± ∞ ¹        ~ (p=1.000 n=1) ²
UnmarshalVertexLog-10      221.5n ± ∞ ¹   187.3n ± ∞ ¹        ~ (p=1.000 n=1) ²
geomean                    337.7n         220.8n        -34.62%
¹ need >= 6 samples for confidence interval at level 0.95
² need >= 4 samples to detect a difference at alpha level 0.05

pkg: github.com/moby/buildkit/cache/contenthash
                         │ ../gogo.txt  │            ../vtproto.txt            │
                         │    sec/op    │    sec/op     vs base                │
MarshalCacheRecords-10     341.0n ± ∞ ¹   346.3n ± ∞ ¹       ~ (p=1.000 n=1) ²
UnmarshalCacheRecords-10   899.3n ± ∞ ¹   952.7n ± ∞ ¹       ~ (p=1.000 n=1) ²
geomean                    553.8n         574.4n        +3.72%
¹ need >= 6 samples for confidence interval at level 0.95
² need >= 4 samples to detect a difference at alpha level 0.05
```

There seems to be a decent regression in speed for checksumming. The marshaling for the progress stream doesn't seem to be as significant.