@jsternberg 2024-09-24
Separating the vtproto component of this into a separate PR.