@jsternberg 2024-09-25
We did some investigation of memory usage when transferring a large context and found some interesting stuff. The big thing we found is that a certain optimization utilized [here](https://github.com/tonistiigi/fsutil/blob/536d6131a9d8df76af194d8946a81dbc5bc6fec1/receive.go#L200-L203) does nothing and doesn't help with memory usage. This is because of the default grpc codec and is seen across different implementations.

The default codec will always call `Reset` on the underlying message. This can be seen [here](https://github.com/moby/buildkit/blob/856ae24aa74793c122533d6bfdeffee26cb6e80b/vendor/google.golang.org/protobuf/proto/decode.go#L87-L89). The `Merge` is set to `false` by default so the reset always happens. The implementation of reset in gogo/protobuf [removes the saved buffer](https://github.com/moby/buildkit/blob/856ae24aa74793c122533d6bfdeffee26cb6e80b/vendor/github.com/tonistiigi/fsutil/types/wire.pb.go#L66) so we always allocate memory for the data we are transferring.

This behavior is consistent with the standard go implementation which does the same thing, but it seems the standard implementation benefits a little bit from utilizing a v2 version of the codec that incorporates a memory pool. Still, it allocates a lot and that dominates the performance.

vtproto also does the same thing if you do the same thing, but vtproto can generate its own reset function that's useful for memory pools and we can also customize the codec to remove the reset if that makes sense. Either way, I haven't seen a regression in the performance so far but I think we can like do much more with vtproto if we dive into things more thoroughly.